refactor(portal): split problem panel actions
